tests/submission/W3C/domhtml/HTMLAnchorElement13.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - Anchor</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/editing.html#dom-blur" rel="help">
<link rel="author" title="Curt Arnold">
</head>
<body>
<p>
<a accesskey="g" charset="US-ASCII" coords="0,0,100,100" dir="LTR" href="/support/images/yellow-rectangle.png" hreflang="en" id="Anchor" name="Anchor" rel="GLOSSARY" rev="STYLESHEET" shape="rect" tabindex="22" type="image/png">View Submit Button</a>
</p>


<script>
var domtest = async_test("HTMLAnchorElement.blur should surrender input focus.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("a");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      testNode.blur();
      
    domtest.done();
}));

</script>
<p>HTMLAnchorElement.blur should surrender input focus.</p>
<div id="log"></div>
</body></html>
